driver: Use BUG_ON instead of WARN_ON for invalid chandef
When the channel definition is invalid, kernel gives a warning
and prints a stack trace. This warning is not of much help to identify
the reason why the chandef is invalid. For better debugging and to
collect dumps for knowing the complete state of the system, replace
WARN_ON with BUG_ON to force crash the system when the chendef is
invalid.

Change-Id: I5d5100bd125a0c89dc3adbdd06597fa7cb5a2bf2
Signed-off-by: Ashishkumar Vijayakumar <ashivija@qti.qualcomm.com>
Signed-off-by: Kaviarasu Sanjeevi <kaviaras@qti.qualcomm.com>