net: PCS: Check and recover pcs link after serdes reset
Currently we dont have reset domain for serdes, so there is no
way for kernel space to reset the serdes which is programmed in the
SCMI platform. With this flexibility to reset serdes from emac
driver added it is good to check for XPCS link status.

After serdes is reset from the emac block, according to programming
guidelines, it is required to check the XPCS link to ensure no
BLK_ERR, BLOCK_LOCK errors and RLU failures have occurred. Retry
if required by resetting the PCS block to ensure group alignment.

Change-Id: I7f3ac61a47afebb7ff488625e188222b92cb94c1
Signed-off-by: Abhishek Chauhan <quic_abchauha@quicinc.com>