phy: qcom-snps-eusb2: Use ref_clk_src for reference clock frequency
Update the driver to use ref_clk_src for determining the reference clock
frequency instead of ref_clk. Add support to acquire the ref_clk_src
clock resource in the probe function, and handle errors if it is not
available. This ensures the correct clock source is used for PHY
initialization and improves compatibility with platforms where the
reference clock is provided via ref_clk_src.

Change-Id: I5a0c8373eeabde6ee73790fda5f734c1ed9dcb4c
Signed-off-by: Pratham Pratap <quic_ppratap@quicinc.com>