FROMLIST: arm64: dts: qcom: Add EL2 overlay for sa8775p
All the Lemans IOT variants boards are using Gunyah hypervisor which
means that, so far, Linux-based OS could only boot in EL1 on those
devices.  However, it is possible for us to boot Linux at EL2 on these
devices [1].

[1]
https://docs.qualcomm.com/bundle/publicresource/topics/
80-70020-4/boot-developer-touchpoints.html#uefi

When running under Gunyah, remote processor firmware IOMMU streams is
controlled by the Gunyah however when Linux take ownership of it in EL2,
It need to configure it properly to use remote processor.

Add a EL2-specific DT overlay and apply it to Lemans IOT variant
devices to create -el2.dtb for each of them alongside "normal" dtb.

Change-Id: I933b2fd4783196bcb641acc25f3c0c1742715815
Link: https://lore.kernel.org/lkml/20251013-kvm_rprocv5-v5-13-d609ed766061@oss.qualcomm.com/
Signed-off-by: Mukesh Ojha <mukesh.ojha@oss.qualcomm.com>