FROMLIST: arm64: dts: qcom: Add support for QCS615 Talos evk board
Introduce the device tree support for the QCS615-based evk
platform, which follows the SMARC (Smart Mobility ARChitecture)
standard. The platform is composed of two main hardware
components: the qcs615-evk-som and the qcs615-evk carrier board.

The qcs615-evk-som is a compact System on Module that integrates the
QCS615 SoC, PMIC, and essential GPIO connectivity. It follows the
SMARC standard, which defines a modular form factor allowing the SoM
to be paired with different carrier boards for varied applications.

The qcs615-evk is one such carrier board, designed for evaluation
and development purposes. It provides additional peripherals
such as UART, USB, and other interfaces to enable rapid
prototyping and hardware bring-up.

This initial device tree provides the basic configuration needed
to boot the platform to a UART shell. Further patches will extend
support for additional peripherals and subsystems.

The initial device tree includes basic support for:

- CPU and memory

- UART

- GPIOs

- Regulators

- PMIC

- Early console

- AT24MAC602 EEPROM

- MCP2515 SPI to CAN

QCS615 evk uses a Quectel AF68E WiFi/BT module (PCIe for
WiFi and UART for Bluetooth), which is different from the RIDE
platform. Plan to enable these in a follow-up patch series.

Change-Id: I19977934bff8f287b6e9702ae89554aa11b3ebe9
Link: https://lore.kernel.org/all/20251014120223.1914790-3-tessolveupstream@gmail.com/
Signed-off-by: Jie Gan <jie.gan@oss.qualcomm.com>