PENDING: net: stmmac:  Add device presence check in TX flush path
A race condition can occur where the transmit path
is executed after the network device has been
unregistered or is in the process of removal. This
may lead to a kernel warning or other memory corruption
issues.

Add a `netif_device_present()` check at the beginning of
`stmmac_flush_tx_descriptors()`. If the device is not
present, trigger a `WARN_ON()` to generate a stack trace
for debugging and return immediately to prevent further
access to hardware. This helps diagnose the race condition
without crashing the system.

Change-Id: I4c101d0847db97e941d792abf68b559852c36371
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>