usb: dwc3: msm: Enable EDP/EDM IRQs to detect soft disconnect during suspend
If a USB device disconnects while the bus is in a suspended state, the
host may fail to detect the disconnect event, resulting in an
inconsistent or unrecognized state.

To address this, configure both EDP and EDM interrupts when the host
issues a suspend. This ensures that any soft disconnect event will
trigger an interrupt (via EDP or EDM going high), allowing the host to
correctly interpret it as a disconnect.

Change-Id: Id23bd9bc48940794a1ee4b0b66c64d576707493a
Signed-off-by: Udipto Goswami <udipto.goswami@oss.qualcomm.com>