BACKPORT: Bluetooth: btusb: QCA: Support downloading custom-made firmwares
There are custom-made firmwares based on board ID for a given QCA BT
chip sometimes, and they are different with existing firmwares and put
in a separate subdirectory to avoid conflict, for example:
QCA2066, as a variant of WCN6855, has firmwares under 'qca/QCA2066/'
of linux-firmware repository.

Support downloading custom-made firmwares based on a table newly added.

Git-commit: a3f9f6dd047af711341a2367a8b08a3ade31438d
Git-repo:
https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/.

Change-Id: I5d28b6eef7404ed6695b744bc504a1313527a5f6
Signed-off-by: Zijun Hu <zijun.hu@oss.qualcomm.com>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
Signed-off-by: Shuai Zhang <quic_shuaz@quicinc.com>