PENDING: driver: bluetooth: btusb: add custom-made firmwares
There are custom-made firmwares based on board ID for a given QCA BT
chip sometimes, and they are different with existing firmwares and put
in a separate subdirectory to avoid conflict, for example:
QCA2066, as a variant of WCN6855, has firmwares under 'qca/QCA2066/'
of linux-firmware repository.

Support downloading custom-made firmwares based on a table newly added.

Change-Id: I85fd24547203077e9f6604a82de6eff79b11d473
Signed-off-by: Shuai Zhang <quic_shuaz@quicinc.com>