kleaf-scripts: modules_register: Add field for library dependencies
Currently, the dependencies fed into the module registry are filtered
against the list of modules being compiled. This does not allow for
other types of dependencies that modules may have, like ddk_library()
rules. To rectify this, add a library dependencies parameter to the
module register function to allow for additional types of dependencies
to be specified.

Change-Id: Id524e32a0fc41c909a1c7c817c003f37578dbb7e
Signed-off-by: Melody Olvera <melody.olvera@oss.qualcomm.com>