qpace: zram: Go on handling event ring if completion interrupt missed
There is one real corner case showing compression completion interrupt
was missed. In such a case we should be able to just directly check HW
event ring and process the ED/completion. No recovery needs to be
performed.

Implement this by waiting for the completion interrupt with a timeout.
Regardless of whether it times out, then poll the last ED to be
consumed. Last polling both works around missing interrupts and checks
for spurious interrupts.

Change-Id: I4f7f5ef314e450185ad5f2238b2065b7e707ed86
Signed-off-by: Zhenhua Huang <quic_zhenhuah@quicinc.com>