soc: qcom: gic_intr_routing: Handle defective cpu case
There are certain scenarios where logical CPU is not equal to the
physical cpu in HLOS. In this case, correct mpidr of the cpu that is
referred by logical cpu number is to be used.

Get the physical cpu number based on logical cpu and use corresponding
physical cpu number to get the mpidr through phandle.

Change-Id: Ied6e56a3b45315fa1e6a99343896e69bc08a5f52
Signed-off-by: Lakshmi Sowjanya D <laksd@qti.qualcomm.com>