pci: msm: create a de-enumerate api
SD Express endpoint requires hotplug detection, there is a detection
driver which monitors card detect gpio to know card is inserted or
removed.

As part of card removal, we need to make sure we are de-registering
from the pci platform drivers and deiniting clks etc.

For this purpose, created this api for de-registering.

Also, protect cfg_access with a spinlock in msm_pcie_enumerate() to
prevent race conditions.

Change-Id: If67e3b34f0efc1959624f3d43d1de0ee6d09101a
Signed-off-by: Krishna Chaitanya Chundru <krichai@qti.qualcomm.com>
Signed-off-by: Matthew Leung <mattleun@qti.qualcomm.com>
Signed-off-by: Paras Sharma <parass@qti.qualcomm.com>