coresight: byte-cntr: Fix the issue where ODL stops receiving data
During user operations, if certain actions interfere with the
byte_cntr interrupt, ODL may continue running but fail to receive
trace data. This change introduces a timeout while waiting for the
byte_cntr interrupt. If the driver does not receive the interrupt
within 5 seconds, it calculates the current rwp offset to check if
there is data pending for transfer. If data exists, the driver
invokes the API to push the remaining data to user space,
preventing data loss while ODL is running.

Call the function "tmc_etr_read_bytes" if the requirement size
is greater than the block size when the byte counter IRQ doesn't
come.

Change-Id: I3a660fc10117f185067a90deba7ee285204a673c
Signed-off-by: Tao Zhang <quic_taozha@quicinc.com>