FROMLIST: iommu/arm-smmu-v3-kvm: Shadow the command queue
At boot allocate a command queue per SMMU which is used as a shadow
by the hypervisor.

The command queue size is 64K which is more than enough, as the
hypervisor would consume all the entries per a command queue prod
write, which means it can handle up to 4096 at a time.

Then, the host command queue needs to be pinned in a shared state, so
it can't be donated to VMs, and avoid tricking the hypervisor into
accessing them. This is done each time the command queue is enabled,
and undone each time the command queue is disabled.
The hypervisor won’t access the host command queue when it is disabled
from the host.

Bug: 461757236
Bug: 458241298
Change-Id: If6fed351c6cf112b8faf46734e257ea4d692094a
Link: https://patch.msgid.link/20251117184815.1027271-20-smostafa@google.com
Git-repo: https://android.googlesource.com/kernel/common
Patch-mainline: Not yet mainlined
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 7a1a588fea7925de2220014cfc7e863f57245587)
Signed-off-by: Sricharan Ramabadhran <srichara@qti.qualcomm.com>
Signed-off-by: Maheshwar Sakaray Lambade <msakaray@qti.qualcomm.com>