drivers:iio:stm:imu:st_asm330lhhx: add support to WAKE_THS_W
When setting wake-up detection thresholds below an LSB, there
is a risk that the threshold register will be written to zero,
causing continuous wake-up events to be generated. Using the
WAKE_THS_W bit, it is possible to use an LSB four times lower,
allowing for jobs lower than the threshold expected when
WAKE_THS_W is set to the default.

Change-Id: I8876949581f1b0febec9309137654e5fd6b65f9a
Signed-off-by: Saurav Shah <sauravs@qti.qualcomm.com>