firmware: qcom: si-core: Use #ifdef macros for preparing mem objects
Use compile-time macros to determine whether a memory object should be
prepared via the SHM bridge or the FFA ABI. This approach addresses
scenarios where the QTEE version on the chip supports FFA, but the
si-core driver still relies on the SHM bridge. In such cases, QTEE depends
on si-core to indicate the appropriate mapping operation to use on its
side.

Change-Id: I8bc99027242d880e1050e3bc95350fdf1239274d
Signed-off-by: Harshal Dev <quic_hdev@quicinc.com>
Signed-off-by: Kirtan Shah <kirtshah@qti.qualcomm.com>