stmmac: Make stmmac_rx_vlan a no-op function
With SPH enabled, dvlan packet HL is being set to 14 instead of 22.
Post sw vlan tag stripping HL is less than 14 leading to crash in
skbuff handing.

Removing sw vlan tag stripping logic as it is not required
for vlan datapath to work and causing HL mismatch when SPH
is enabled.

Change-Id: Ib8d91a2bbf28d70b60f7357fb8e7669c79b7dcfa
Signed-off-by: Veena Sri Vallala <quic_vsri@quicinc.com>