PENDING: net: PCS: Check for AN status before posting link up
Currently after serdes reset we check for just link status
bit in case of 10GbaseR and 5GbaseR cases which does not
support Auto-negotiation.

In case of USXGMII, Auto-negotiation is supported. We need to check
for AN status before we post link as up to the MAC.

Change-Id: I690c99c60821166177fa1cff1916ac3ceb4a8f0e
Signed-off-by: Abhishek Chauhan <quic_abchauha@quicinc.com>
(cherry picked from commit 17b30a37b6a76fe70d450353a4f820a9216af261)