zram: qpace: Add suspend callback
Qpace cannot accept new requests while it is in the suspend flow.
Although a wakelock is held when queuing requests, unfreezable
tasks may still enqueue requests that bypass the wakelock check.
To prevent this, set the qpace_suspended flag in the suspend
callback to block further queuing. If any request is detected during
the suspend callback, abort the suspend sequence. The flag is reset
in the resume callback.
A mutex lock is used to protect both the active rings and the
qpace_suspended flag.

Change-Id: Ifee9673f822bd0aef0cb6e0f2ff43bce481aa4c6
Signed-off-by: Zhenhua Huang <zhenhua.huang@oss.qualcomm.com>