net: stmmac: address non-LPI resume failures properly
The Synopsys Designware GMAC core databook requires all clocks to be
active in order to complete software reset, which we perform during
resume.

However, IEEE 802.3 allows a PHY to stop its clocks when placed in
low-power mode, which happens when the system is suspended and WoL
is not enabled.

As an attempt to work around this, commit 36d18b5664ef ("net: stmmac:
start phylink instance before stmmac_hw_setup()") started phylink
early, but this has the side effect that the mac_link_up() method may
be called before or during the initialisation of GMAC hardware.

We also have the socfpga glue driver directly calling phy_resume()
also as an attempt to work around this.

In a previous commit, phylink_prepare_resume() has been introduced
to give MAC drivers a way to ensure that the PHY is resumed prior to
their initialisation of their MAC hardware. This commit adds the call,
and moves the phylink_resume() call back to where it should be before
the aforementioned commit.

The full upstream change depends on a phylink core
fix (commit 367f1854d442, "net: phylink: ...") which is not yet
present in the ACK / Google kernel branch used by Nord GVM.
On this branch, phylink is built-in as a static module from ACK,
so calling phylink_prepare_resume() without that core fix will
lead to compile error.

For now, only the stmmac-side changes that restructure the resume
flow are brought in, while the phylink_prepare_resume(priv->phylink);
call is intentionally omitted. We will raise a ticket with Google to
backport the upstream phylink fix (367f1854d442) into their branch,
we will follow up with a separate change to enable the
phylink_prepare_resume() call from the stmmac resume path.

Signed-off-by: Russell King (Oracle) <rmk+kernel@armlinux.org.uk>
Link: https://patch.msgid.link/E1tvO6a-008Vjh-FG@rmk-PC.armlinux.org.uk
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Change-Id: I021b1f513af347298350b5c413aff5a7e23292bd
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>