pkvm: qcom: Add SMC filter module prototype
Initial version of the SMC filter module. The module provides
filtering functionality for SMC calls originating from the host in
pKVM environment. It intercepts SMC calls and controls which calls
are permitted.

If using the extended SCM ABI, with x5 containing a pointer to the
arguments buffer, the memory area is shared with the hypervisor and
pinned for the host for the duration of the SMC call. The SMC is then
forwarded manually by this module, which then unwinds the memory
sharing and pinning.

Change-Id: I0a12ca404686d18067a1c3599b8b81f1403e65a7
Co-developed-by: Carl van Schaik <cvanscha@qti.qualcomm.com>
Signed-off-by: Carl van Schaik <cvanscha@qti.qualcomm.com>
Signed-off-by: Yair Kazarinov <ykazarin@qti.qualcomm.com>
Signed-off-by: Maheshwar Sakaray Lambade <msakaray@qti.qualcomm.com>