usb: dwc3-msm-core: Fix list corruption in perf vote work
Currently in msm_dwc3_perf_vote_work() the function
msm_dwc3_perf_vote_update() would get called irrespective of weather a
qos request is active or not which is a potential risk causing list
corruption. This indicates that even when request isn't active the
function tries to update a stale entry of the request.

Fix this by adding a check of request active before executing the
function.

Change-Id: Ia212fccd2a6e3cb86b50651426eb49e72d20a139
Signed-off-by: Udipto Goswami <udipto.goswami@oss.qualcomm.com>