PENDING: net: stmmac: Avoid access of register after suspend
Xmit of packet and suspending of device can happen in parallel
and in different context. It can happen that packet can arrive from
the network stack while the suspend is in progress. In such cases
we need to check the netdev device presence before we try to access
the tail pointer register to avoid access of emac register when the
clocks are disabled.

Change-Id: I4c4b9ca71a1160bfbccadf583b4e02237e2f3453
Signed-off-by: Abhishek Chauhan <abchauha@qti.qualcomm.com>