PENDING: net: stmmac: Fix safety interrupt registration in multi-IRQ mode
The IRQF_SHARED flag was missing in case of safety interrupt when
Multi IRQ mode is enabled, which is required as this interrupt is
shared with the PCS device, additionally the snprintf buffer size
for the interrupt name was incorrect, leading to an empty name string.
This patch corrects the buffer size and adds the necessary flag
to ensure proper registration.

Change-Id: I8e5d27b9baf445476ead5312a0683421c5e7c797
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>