net: pcs: Update the XPCS sequence and add recovery when link fails
In order to resolve the CRC issue, we update the XPCS sequence and
add a recovery mechanism when the link fails to ensure the XPCS
link comes up.

1. Reset SerDes before bringing up the XPCS link.
2. Bring up the XPCS link.
3. If the XPCS link fails:
	1. Reset SerDes again.
	2. Reset XPCS and check the link status.
	3. If the XPCS link still fails, goto #3.

RLU(Rx Link Up) bit is a Latched low bit, the default value is 1.
When this bit is cleared, it does not change until the host reads
the register. so if we need to know the current status, we need to
read this register twice.

Change-Id: I21dc8f16ed1df7b1b68ccdef7376046f31aa3edf
Signed-off-by: Mei Xie <meixie@qti.qualcomm.com>