soc-repo: drivers: Replace PM notifier warning with pr_debug()
The current hibernation PM notifier uses WARN_ONCE() for unexpected
notifier events. It can give the impression that this driver is
failing and fills thelogs with unnecessary stack traces for
non-hibernation PM calls.

Replace this with pr_debug() so that unexpected notifier values are
still visible to developers when debugging, but do not pollute logs
or raise unnecessary alarm in non-hibernation PM calls.

Change-Id: If1039658cbf62f0489fbe45976f2614ff9ba33ba
Signed-off-by: NEERAJ SHARMA <neesha@qti.qualcomm.com>